refactor: notes

+++ src/notes/notes.repository.ts
@@ -1,54 +1,31 @@
 import { PrismaService } from '@/prisma/prisma.service';
 import { Injectable } from '@nestjs/common';
-import { NotesDTO } from './dto/notes.dto';
+import { NotesDto } from '@/notes/dto/notes.dto';
 
 @Injectable()
 export class NotesRepository {
   constructor(private readonly prisma: PrismaService) {}
 
-  createNotes(noteDTO: NotesDTO, userId: number) {
-    return this.prisma.note.create({
-      data: {
-        ...noteDTO,
-        userId,
-      },
-    });
+  createNotes(noteDTO: NotesDto, userId: number) {
+    return this.prisma.note.create({ data: { ...noteDTO, userId } });
   }
 
   getNoteByTitleAndUserId(title: string, userId: number) {
-    return this.prisma.note.findUnique({
-      where: {
-        userId_title: {
-          userId,
-          title,
-        },
-      },
-    });
+    return this.prisma.note.findUnique({ where: { userId_title: { userId, title } } });
   }
 
   findAllNotes(userId: number) {
-    return this.prisma.note.findMany({
-      where: { userId: userId },
-    });
+    return this.prisma.note.findMany({ where: { userId: userId } });
   }
 
   findNoteById(id: number) {
-    return this.prisma.note.findUnique({
-      where: { id },
-    });
+    return this.prisma.note.findUnique({ where: { id } });
   }
 
   deleteNote(id: number) {
-    return this.prisma.note.delete({
-      where: { id },
-    });
+    return this.prisma.note.delete({ where: { id } });
   }
 
   deleteAllNotes(userId: number) {
-    return this.prisma.note.deleteMany({
-      where: {
-        userId,
-      },
-    });
-  }
+    return this.prisma.note.deleteMany({ where: { userId } }) }
 }
